+++ syscall/samples/module/create_module.c
@@ -1,0 +1,17 @@
+/**
+ * caddr_t create_module(const char *name, size_t size);
+ *
+ * This system call is present only before Linux 2.6.
+ */
+#include <stdio.h>
+#include <stdlib.h>
+#include <errno.h>
+#include <linux/module.h>
+
+
+int main(int argc, char *argv[])
+{
+	/* TODO */
+	return 0;
+}
+